refactor: move system-logic from model to controller
