+++ core/src/main/kotlin/com/tdt4240/group3/Hexa_Battle.kt
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.g2d.BitmapFont
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
 import com.badlogic.ashley.core.Engine
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer
-import com.tdt4240.group3.model.systems.PlayerSystem
+import com.tdt4240.group3.controller.systems.PlayerSystem
 import com.tdt4240.group3.screens.HowToPlayScreen
 import com.tdt4240.group3.view.systems.View
 import ktx.app.KtxGame
