+++ core/src/main/kotlin/com/tdt4240/group3/controller/TroopCreationController.kt
@@ -1,8 +1,7 @@
 package com.tdt4240.group3.controller
 
-import com.tdt4240.group3.model.components.TeamComponent
-import com.tdt4240.group3.model.systems.TroopCreationSystem
-import com.tdt4240.group3.model.systems.TurnSystem
+import com.tdt4240.group3.controller.systems.TroopCreationSystem
+import com.tdt4240.group3.controller.systems.TurnSystem
 
 class TroopCreationController(
     private val troopCreationSystem: TroopCreationSystem,
