+++ core/src/main/kotlin/com/tdt4240/group3/controller/systems/PlayerSystem.kt
@@ -1,4 +1,4 @@
-package com.tdt4240.group3.model.systems
+package com.tdt4240.group3.controller.systems
 
 import com.tdt4240.group3.model.components.PlayerComponent
 import com.badlogic.ashley.systems.IteratingSystem
